feat: render exemple usage preview for each component
