+++ content/docs/button.mdx
@@ -3,27 +3,3 @@ title: Button
 description: Buttons allow users to take actions, and make choices, with a single tap. They are typically placed throughout your UI, in places like dialogs, forms, cards, and toolbars.
 date: 2022-03-11
 ---
-
-# I hate buttons
-
-## Very much
-
-### They are so annoying
-
-### I hate them
-
-- button
-
-1. button
-2. button
-3. button
-
-- _italic_
-- **bold**
-- `code`
-
-```bash
-bunx shadcn@latest add https://shuip.xyz/r/button.signout.json
-bunx shadcn@latest add https://shuip.xyz/r/button.submit.json
-bunx shadcn@latest add https://shuip.xyz/r/button.theme.json
-```
